+++ src/photod/stats.py
